Changed EntryImportModel: added APT and RAW image fields; made the multiple file select appendable and unique. Added import logic. Updated EntryEditor to display information and started working on making it editable.

+++ Shared/JournalEntry/Sources/CaptureSequenceList.swift
@@ -36,7 +36,6 @@ struct CaptureSequenceList: Codable, Hashable {
     let negativeDec: Bool
     let captureSequences: [CaptureSequence]
 
-
     enum CodingKeys: String, CodingKey {
         case targetName = "TargetName"
         case mode = "Mode"
@@ -70,7 +69,7 @@ struct CaptureSequenceList: Codable, Hashable {
     
     struct CaptureSequence: Codable, Hashable {
         let enabled: Bool
-        let exposureTime: Int
+        let exposureTime: Double
         let imageType: String
         let filterType: FilterType
         let binning: Binning
